Debug Options and Paragraphs
Wrote some debug buttons for death and insanity, also did quite a bit of work to hide parts of the UI that needs to not be visible. Got the investigator stats to display and be able to be updated. Did some work on the passage of time in the game, but I think I will still rework this.....not sure yet.  Tracking time using a calendar would be tricky, I think.

Started work on Paragraph advancement.....this would probably be much easier in another language or with a library, but this is all I know at this point.

Tricky part would be only advancing the day once through a paragraph....guess I could use a boolean to decide if it has happened already, just don't know if that is the best way forward.